Remove all but most recent HHVM 3
As per this announcement (https://hhvm.com/blog/2018/12/17/hhvm-3.30.html), Facebook are deprecating PHP support in HHVM. 3.30 is the last LTS version. As such, I've removed all old versions from the tests. As soon as docopt breaks with 3.30, I will probably bump the major version of docopt.php, remove all vestiges of support for HHVM, and direct users to install the old version if they need to keep using it.